pyproject.toml
[project]
name = "checkbox-ua"
version = "0.1.0"
description = ""
authors = [
    {name = "Serhii Tolmachov",email = "cvaizencva@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "bcrypt (>=4.3.0,<5.0.0)",
    "jinja2 (>=3.1.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.7"

